requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2024.8.30
distro==1.9.0
EbookLib==0.18
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
lxml==5.3.0
openai==1.56.1
pydantic==2.10.3
pydantic_core==2.27.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
